Move DB connection string
