+++ TescEvents/Entities/RepositoryContext.cs
@@ -1,5 +1,6 @@
 using Microsoft.EntityFrameworkCore;
 using TescEvents.Models;
+using TescEvents.Utilities;
 
 namespace TescEvents.Entities; 
 
@@ -7,12 +8,7 @@ public class RepositoryContext : DbContext {
     public RepositoryContext(DbContextOptions options) : base(options) { }
 
     protected override void OnConfiguring(DbContextOptionsBuilder options) {
-        var host = Environment.GetEnvironmentVariable("DB_HOST");
-        var port = Environment.GetEnvironmentVariable("DB_PORT");
-        var database = Environment.GetEnvironmentVariable("DB_DATABASE");
-        var user = Environment.GetEnvironmentVariable("DB_USER");
-        var pass = Environment.GetEnvironmentVariable("DB_PASS");
-        options.UseNpgsql($"Host={host};Port={port};Database={database};Username={user};Password={pass}");
+        options.UseNpgsql(AppSettings.ConnectionString);
     }
     
     public DbSet<Event>? Events { get; set; }
