+++ TescEvents/Program.cs
@@ -12,22 +12,13 @@ DotEnv.Load(dotenv);
 
 builder.Services.AddControllers();
 builder.Services.AddDbContext<RepositoryContext>(options => 
-                                                     options.UseNpgsql(builder.Configuration.GetConnectionString("RepositoryContext")));
-// Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
-builder.Services.AddEndpointsApiExplorer();
-builder.Services.AddSwaggerGen();
+                                                     options.UseNpgsql(AppSettings.ConnectionString));
 
 builder.Services.AddScoped<IEventRepository, EventRepository>();
     
 builder.Configuration.AddEnvironmentVariables();
 
 var app = builder.Build();
-
-// Configure the HTTP request pipeline.
-if (app.Environment.IsDevelopment()) {
-    app.UseSwagger();
-    app.UseSwaggerUI();
-}
 
 app.UseHttpsRedirection();
